refactor: inner class to upper level
